Update test.py and add Eve REST framework
The test.py should now contain a sample implement to pull rdl data.
We are currently pulling copy 4880 and we'll be mapping out the rest
of the Restful routes. After this we'll move this to production and
automate the process. This will allow us to finally get to work on the
front-end of the application